Makefile.inc: specify package version
Due to a new build environment we won't use a spec file from the
same repository as the sources, so we can't rely on .spec file
to hold the version of the package. Therefore, let's specify
package version (and optional RELEASE which should usually be empty)
in Makefile.inc which is included into other Makefiles.

lib/Makefile determines library version from the package version the
same way as before commit 5691b39.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>