normalize_image_name(): fix for corner case
Chances to hit this bug are pretty slim, but I just did it.

For the following input parameters:

 basedir="/vz/private/104/root.hdd"
 image="/vz/private/104/root.hdd.{86be108e-d30e-4054-b6d4-412d4d0c2ce6}"

function errorneously returned

 {86be108e-d30e-4054-b6d4-412d4d0c2ce6}

Apparently, we need to add a check for '/' right after basedir. Note
we do this check right after strncmp which returned 0 -- this guaranees
that image is at least as long as basedir.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>