ploop_replace_image(): optimization for keep_name
In case KEEP_NAME flag is set, we need to rename the new image file
to the current (old) one. But as in-kernel ploop driver doesn't allow
to remove (or rename(), as it comes out) a delta file which is in use
unless its hardlink count is > 1, we have to make a trick here:
 - create a hardlink to the file
 - rename it
 - remove the hardlink

It just occured to me we can avoid doing this trick if the file already
has more than 1 hardlink. This is what this patch does.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>