ploop replace: implement offline replace
An offline replace is just a way to replace a file name in
DiskDescriptor.xml with some other file name. In this mode,
same checks are performed in a new delta. Flag --keep-name
is also supported, in such case ploop replace works like
"mv -f" but with bells and whistles (all the checks etc.)

Signed-off-by: Kir Kolyshkin <kir@openvz.org>