ploop_get_top_delta_fname(): use ploop_lock(), return SYSEXIT_*
A problem with ploop_get_top_delta_fname() is it can only be used
when a DiskDescriptor.xml is read into memory. For example, the
following code won't work:

	struct ploop_disk_images_data *di;
	char name[PATH_MAX];

	ret = ploop_open_dd(&di, argv[0]);
	if (ret)
		return ret;
	ret = ploop_get_top_delta_fname(di, name, sizeof(name));
	if (ret)
		return ret;

	printf("top delta: %s", name);

There are three problems:

1 It's not working, we'll have

	Error in ploop_get_top_delta_fname (ploop.c:1192): Can't find image by uuid (null)

2 We will get -1 instead of proper SYSEXIT_* code.

3 If someone will create a snapshot in parallel with us, we'll get old
top_delta name.

This patch fixes #1 and #2 and partially #3. Note #3 is still not fully
fixed, as someone can create snapshot between we call for
top_delta_fname and use it. We need explicit locking for this case.

NOTE1 this patch changes the return values from the function.
If the code calling this was checking for zero or non-zero,
it's fine. If it was explicitly checking for -1 or < 0, it
needs to be fixed.

NOTE2 this is not the only place in the code that is broken since
introducing new locking scheme, as I pointed out in commits
2b80e7b, 63fa44b, and 4307a47. It's a shame it's still not fixed.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>