Add pkgconfig file
This makes it possible to use pkg-config when building stuff that
requires ploop library. For example:

	# pkg-config --libs --static ploop
	-lploop -lrt -lpthread -lxml2 -lz -llzma -lm

Note it shows libs that are not explicitly listed, but are
needed to statically link with libxml2 (they are of course
taken from libxml-2.0.pc).

NOTE: ploop.spec need to include this file to ploop-devel package.
As ploop.spec file is gone from this repo and I can't find it on
src.openvz.org, I can't patch it in this commit.

So, here is a change required to ploop.spec:

	diff --git a/ploop.spec b/ploop.spec
	index 65732a5..f655374 100644
	--- a/ploop.spec
	+++ b/ploop.spec
	@@ -78,6 +78,7 @@ fi
	 %package devel
	 Summary: Headers for development with ploop library
	 Group: Applications/System
	+Requires: pkgconfig
	 %description devel
	 Headers and a static version of ploop library

	@@ -85,6 +86,7 @@ Headers and a static version of ploop library
	 %defattr(-,root,root)
	 %dir %{_incdir}
	 %attr(644,root,root) %{_libdir}/libploop.a
	+%{_libdir}/pkgconfig/ploop.pc
	 %{_libdir}/libploop.so
	 %attr(644,root,root) %{_incdir}/libploop.h
	 %attr(644,root,root) %{_incdir}/ploop_if.h

Signed-off-by: Kir Kolyshkin <kir@openvz.org>