do_ioctl(): implement exponential wait for retry
I was looking into ploop umount delay, and found out it waits for 1
second between retries.

Maxim Patlasov suggested to start with a smaller delay (like 0.01 s) and
increase sleep intervals in an exponential fashion.

With this patch, sleep intervals between retries are doubled each
time. It starts with 0.01s and then increases to max of 0.5s per
interval. Total time is about 60 seconds.

Also, this commit fixes the case of unnecessary sleep
after the last iteration.

While at it, provide more detailed comment why we need to do it.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>