balloon.c: no need to include libgen.h
libgen.h only provides POSIX version of basename()
and dirname(). Neither is used from balloon.c.

In fact it can be dangerous to include libgen.h,
as GNU basename() can be replaced with POSIX basename()
and we might rely on GNU version. As neither is used
from balloon.c it's not a problem this time.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>