ploop_mount_fs(): rectify wrt remount
There are a few problems with the current implementation
of ploop_mount_fs(), such as:

1. The two-step mount is supposed to only add ballon_ino argument
   during the remount. Currently, remount also adds quota arguments.

2. An error from stat() syscall is silently ignored.

3. Even if there is no balloon, i.e. when called from
   create_balloon_file(), a remount is still happening.

This patch fixes all these issues. As this code is also called from
create_balloon_file(), when the balloon does not yet exist, we need
to conditionally ignore error from stat() -- so a new parameter,
need_balloon, is added for that purpose.

In addition, the patch improves the error reporting in case of mount()
failure, making the error message unified and more verbose.

Note that this also makes ploop mount fail if balloon file is
not found (or can't be stat'ed). We discussed it with Maxim Patlasov,
the conclusion is we don't support configurations without
a balloon file so it's better to fail early.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>