complete_running_operation(): unexport
The PL_EXT prefix is supposed to be used by functions
that should not be the part of the external library API,
but for some historical reasons they are. Ideally, there
should be no functions marked with this. In the real world,
we'd like to use some low-level library functions that are
not part of the public API from the ploop tool, and this i
the only reason to have PL_EXT.

This is a fix to commit bd23880.

For the original commits adding PL_EXT, see 62f3389 and 85fc03a.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>