ploop_merge_snapshot_by_guid(): improve messages
As we already printing the details of what we are about to merge
(with both UUIDs), there's no need to repeat this again in a simpler
form (only the UUID that was merged and is not available anymore)
when we are done. What we need is a simple success/failure message.

Also:

1 print if a snapshot to be merged is temporary just before the merge;

2 NULLify parent_guid pointer as it is invalid after the call to
  ploop_di_merge_image() which free's the relevant snapshot data.

3 as we print the generic success message, let's also print the generic
  failure message, too, in case of failure.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>