Add functions to resize partition on block device
We try to resize the last partition on the disk.
If you call ploop resize -s 0 /dev/... then in this case
size=0 means that resize partition until the end of the disk.
We use sgdisk tool to restore second GPT header. It's necessary
if we have copied data from original disk to a bigger one.
(Openstack does it. It copies data via 'dd' from a small ploop disk
to a large LVM volume).