Do not use real path for block devices
For LVM devices it's better to use a symbolic links from
/dev/mapper/ directory (Openstack does it).
Because a real path for LVM device looks like /dev/dm-N,
where N is a number.
And if we create a partition on LVM device it will be present like
dm-(N+1). But partition number doesn't match to dm-* number.
For example if you have LVM device dm-2 and you create
partition number 4 it will present as dm-3.
After that you create partition 1 and it will presetn as dm-4.
Also LVM may reuse free numbers of dm-* if you delete a partition.