ploop_init_image(): unable to recreate encrypted image #PSBM-71894
Error: Partition(s) 1 on /dev/ploop59001 have been written, but we have been unable to inform the kernel of the change, probably because it/they are in use.  As a result, the old partition(s) will remain in use.  You should reboot now before making further changes.

Skip partition creation if exists

Signed-off-by: Igor Sukhih <igor@virtuozzo.com>