Do not treat ploop disk as encrypted if keyid is not set #PSBM-140446
Scenario for this problem is rather uncommon, but not impossible.
If ploop image is used by vstorage-iscsi, and VM owner decides to
encrypt the disk, we'll end up with a block device with encrypted
contents. This is fine until you decide to remount it, and ploop
utils start treating this "ploop with encrypted contents" as
"encrypted ploop", encrypted using ploop utility. We must distinguish
those scenarios and require encryption key to treat image as encrypted
ploop.

Signed-off-by: Pavel Vokhmyanin <pvokhmyanin@virtuozzo.com>