ploop: Fix two step mount balloon_ino detection when fs is already mounted, #PSBM-143223
The two step mount first mounts the device without balloon_ino,
then tries to stat the balloon file to obtain the inode number.
And finally it remounts the file system with balloon_ino=number.
In the case  stat fails because the file system is already mounted
and it get EPERM error.

in commit f79a7ae75f9e ("Unable to operate with additional ploop images #PSBM-101605")
handling for the EPERM was added in auto_mount_fs - after the call
to mount_fs to mount the already mounted in a CT device to host but
this time without check for balloon.

Checking for EPERM in auto_mount_fs several have problems:
  - umount can change errno before return
  - logs confusing error messages - about stat and mount
  - does mount/umount/mount

to solve above issues handle EPERM inside the mount_fs function.

Signed-off-by: Alexander Atanasov <alexander.atanasov@virtuozzo.com>