ploop: implement dm-ploop replace functionality, #VSTOR-71011
Replacing deltas in classic ploop and dm-ploop is very different.
Main difference is that since dm-ploop does not expose currently
loaded deltas via sysfs it can not support replace without ddxml
file. We need to query the driver via get_img_name and match files
by name. Other difference is the kernel implementation of the replace
one is done by the driver the other is implemented in device mapper.
When replacing we must take care of flush and lockfs flags of device
mapper to avoid triggering io errors and preserve fs state properly.

https://pmc.acronis.work/browse/VSTOR-71011
Signed-off-by: Alexander Atanasov <alexander.atanasov@virtuozzo.com>