ploop: implement dm-ploop replace delta inside ploop tool, #VSTOR-71011
Detect which version of ploop is used and use respective library call
to replace deltas.

Since dm-ploop requires DiskDescriptor.xml file to be backward
compatible try to detect it in the same directoty as the image
specified for replacement. Only detect if not provided ad argument.

Make some of the error messages a bit more descirptive.

https://pmc.acronis.work/browse/VSTOR-71011
Signed-off-by: Alexander Atanasov <alexander.atanasov@virtuozzo.com>