ploop: clear in use flag when replacing images, #VSTOR-72842
In block ploop kernel manages the in use flag, on start/stop in
device mapper based ploop it must be managed in userspace tool.

When replacing a delta clear in use flag for the old delta, since
it is very likely write to that image to fail in real world scenario
just log an error and continue.

Signed-off-by: Alexander Atanasov <alexander.atanasov@virtuozzo.com>