ploop: reduce defrag tools output when not running on a tty, #PSBM-151232
Both defragmentation utilities e4defrag and xfs_fsr produce an output
tailored for console with lots of escape sequencies - making progress bar.
This confuses the python test tools that collect the output and they timeout.
Redirect output to /dev/null to avoid that.
xfs_fsr starts to log to syslog if not on a terminal but e4defrag does not
have such feature. Both tools do not have quiet mode option.

https://virtuozzo.atlassian.net/browse/PSBM-151232
Signed-off-by: Alexander Atanasov <alexander.atanasov@virtuozzo.com>