ploop: fix crash on umount when device is not available, #VSTOR-76762
On umount a ploop_check is performed on the device and on failure
it is retried with the device suspended but during HA event
it is possible device to be unavailable and the second check to crash
on suspend. Handle gracefuly this situation - do not attempt to
suspend a NULL device.

https://pmc.acronis.work/browse/VSTOR-76762
Signed-off-by: Alexander Atanasov <alexander.atanasov@virtuozzo.com>