ploop: Disable xfs defrag since it only works on files, #PSBM-151232, #PSBM-151557
Disable xfs defrag until it works properly
We want to defragment filesystem and release unused blocks
but xfs_fsr only does defragment files. The problem is that
defragmentation process makes copies of each file and allocates
more clusters instead of releasing free ones.
As a result after the temporary files are removed we observe
x5-8 grow of used space on the undelaying device.

https://virtuozzo.atlassian.net/browse/PSBM-151232
https://virtuozzo.atlassian.net/browse/PSBM-151557
Signed-off-by: Alexander Atanasov <alexander.atanasov@virtuozzo.com>